# Remove answers to fit 512 byte when truncating