@tsenart 2015-10-19
This change set needs tests. It'd be also good to benchmark how it performs. I'm curious :-)