@discordianfish 2015-10-20
@tsenart I started working on tests but wanted to spin up a test cluster so I don't have to hand craft the state.json. Will do that tomorrow.