@tsenart 2015-10-20
@discordianfish: You shouldn't have to do so. I'd like unit tests for this function.