@tsenart 2015-10-20
@discordianfish: If you want, we can pair on it this week.