@discordianfish 2015-11-06
I've added a unit test for the truncate function and some benchmarks. Probably not ideal, so happy for suggestions to improve it. Still need to build better intuition around what/how to test exactly.

Benchmark results:
- Without my changes
```
BenchmarkHandlers	    2000	    927056 ns/op
BenchmarkTruncate	2000000000	         0.30 ns/op
```

- With my changes
```
BenchmarkHandlers	    2000	   1022113 ns/op
BenchmarkTruncate	2000000000	         0.36 ns/op
```