@jdef 2015-11-11
performance of `truncate` degrades significantly as the number of results increases. Instead of using a linear search for a fit, could we use something along the lines of a binary search (O(log n))?