@discordianfish 2015-11-11
@jdef Not sure what could be possibly more efficient in the truncate function than just looping over the answers until they fit the given size. You have some more specific suggestions?
Another way to optimize that would be to check if answers fit when adding/generating them. But that requires some significant change to the code I'd say.