@discordianfish 2015-11-16
@jdef Okay, I've change it to a binary search but now only truncate answers to keep it simple. That's what skydns is doing as well. Performance is much better now:

```
BenchmarkHandlers	    2000	   1049782 ns/op
BenchmarkTruncate	   30000	     54031 ns/op
```