@jdef 2015-11-16
love the performance improvement. left additional review comments.