@jdef 2015-11-17
minor test case comment, otherwise lgtm