@discordianfish 2015-11-17
@jdef Ok cool! Not sure what the process is, can I just merge it now? Or do we need another reviewer?