@jdef 2015-11-17
Safe to merge at this point. @sargun is the owner of this repo now, but we talked yesterday about this PR and he wants this fix in the next release.